# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_Snowitty-Re_CNtunyuan_internal_domain_entity.AuditLog:
    properties:
      action:
        description: 操作：创建、更新、删除、登录等
        type: string
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        description: 操作描述
        type: string
      duration_ms:
        description: 执行时长（毫秒）
        type: integer
      error_message:
        description: 错误信息
        type: string
      id:
        type: string
      module:
        description: 模块：用户管理、案件管理、任务管理等
        type: string
      request_body:
        description: 请求体（敏感信息脱敏）
        type: string
      request_ip:
        description: 请求IP
        type: string
      request_method:
        description: HTTP方法
        type: string
      request_url:
        description: 请求URL
        type: string
      response_body:
        description: 响应体（简化）
        type: string
      response_code:
        description: 响应状态码
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_domain_entity.AuditLogStatus'
        description: 状态
      trace_id:
        description: 追踪ID
        type: string
      type:
        allOf:
        - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_domain_entity.AuditLogType'
        description: 类型
      updated_at:
        type: string
      user_agent:
        description: 用户代理
        type: string
      user_id:
        type: string
      user_role:
        type: string
      username:
        type: string
    type: object
  github_com_Snowitty-Re_CNtunyuan_internal_domain_entity.AuditLogStats:
    properties:
      error_count:
        type: integer
      login_count:
        type: integer
      operation_count:
        type: integer
      today_count:
        type: integer
      total_count:
        type: integer
    type: object
  github_com_Snowitty-Re_CNtunyuan_internal_domain_entity.AuditLogStatus:
    enum:
    - success
    - failure
    type: string
    x-enum-comments:
      AuditLogStatusFailure: 失败
      AuditLogStatusSuccess: 成功
    x-enum-descriptions:
    - 成功
    - 失败
    x-enum-varnames:
    - AuditLogStatusSuccess
    - AuditLogStatusFailure
  github_com_Snowitty-Re_CNtunyuan_internal_domain_entity.AuditLogType:
    enum:
    - login
    - logout
    - create
    - update
    - delete
    - query
    - export
    - upload
    - download
    - other
    type: string
    x-enum-comments:
      AuditLogTypeCreate: 创建
      AuditLogTypeDelete: 删除
      AuditLogTypeDownload: 下载
      AuditLogTypeExport: 导出
      AuditLogTypeLogin: 登录
      AuditLogTypeLogout: 登出
      AuditLogTypeOther: 其他
      AuditLogTypeQuery: 查询
      AuditLogTypeUpdate: 更新
      AuditLogTypeUpload: 上传
    x-enum-descriptions:
    - 登录
    - 登出
    - 创建
    - 更新
    - 删除
    - 查询
    - 导出
    - 上传
    - 下载
    - 其他
    x-enum-varnames:
    - AuditLogTypeLogin
    - AuditLogTypeLogout
    - AuditLogTypeCreate
    - AuditLogTypeUpdate
    - AuditLogTypeDelete
    - AuditLogTypeQuery
    - AuditLogTypeExport
    - AuditLogTypeUpload
    - AuditLogTypeDownload
    - AuditLogTypeOther
...
  github_com_Snowitty-Re_CNtunyuan_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      trace_id:
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  internal_interfaces_http_handler.AuditModuleStatItem:
    description: 模块操作统计项
    properties:
      count:
        description: 操作次数
        example: 42
        type: integer
      module:
        description: 模块名称
        example: 用户管理
        type: string
    type: object
  internal_interfaces_http_handler.AuditPaginatedResult:
    description: 审计日志分页查询结果
    properties:
      items:
        description: 日志列表
        items:
          $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_domain_entity.AuditLog'
        type: array
      page:
        description: 当前页
        type: integer
      page_size:
        description: 每页数量
        type: integer
      total:
        description: 总数
        type: integer
    type: object
  internal_interfaces_http_handler.AuditUserActivityItem:
    description: 用户活动统计项
    properties:
      count:
        description: 操作次数
        example: 15
        type: integer
      date:
        description: 日期
        example: "2024-01-01"
        type: string
    type: object
...
  internal_interfaces_http_handler.CleanupRequest:
    properties:
      days:
        maximum: 365
        minimum: 1
        type: integer
    required:
    - days
    type: object
  internal_interfaces_http_handler.CleanupResponse:
    properties:
      before:
        type: string
      deleted_count:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@cntunyuan.org
    name: CNtunyuan Team
    url: https://github.com/Snowitty-Re/CNtunyuan
  description: 团圆寻亲志愿者系统后端 API 文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/Snowitty-Re/CNtunyuan
  title: 团圆寻亲志愿者系统 API
  version: 1.0.0
paths:
  /audit/cleanup:
    post:
      consumes:
      - application/json
      description: 清理指定天数之前的审计日志，只有超级管理员可以执行此操作
      parameters:
      - description: 清理请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handler.CleanupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_interfaces_http_handler.CleanupResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "403":
          description: 禁止访问，需要超级管理员权限
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 清理旧日志
      tags:
      - 审计日志
  /audit/logs:
    get:
      consumes:
      - application/json
      description: 分页查询审计日志列表，支持按用户、模块、操作类型、状态、时间范围等条件筛选
      parameters:
      - default: 1
        description: 页码，默认1
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 20
        description: 每页数量，默认20
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - description: 用户ID
        in: query
        name: user_id
        type: string
      - description: 模块名称，如：用户管理、案件管理等
        in: query
        name: module
        type: string
      - description: 操作名称，如：创建、更新、删除等
        in: query
        name: action
        type: string
      - description: 日志类型：login/logout/create/update/delete/query/export/upload/download/other
        in: query
        name: type
        type: string
      - description: 状态：success/failure
        in: query
        name: status
        type: string
      - description: 开始时间，格式：2006-01-02
        in: query
        name: start_time
        type: string
      - description: 结束时间，格式：2006-01-02
        in: query
        name: end_time
        type: string
      - description: 关键词搜索
        in: query
        name: keyword
        type: string
      - description: 请求IP地址
        in: query
        name: request_ip
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_interfaces_http_handler.AuditPaginatedResult'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取审计日志列表
      tags:
      - 审计日志
  /audit/logs/{id}:
    get:
      consumes:
      - application/json
      description: 根据审计日志ID获取单条日志详情
      parameters:
      - description: 审计日志ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_domain_entity.AuditLog'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "404":
          description: 日志不存在
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 根据ID获取审计日志
      tags:
      - 审计日志
  /audit/module-stats:
    get:
      consumes:
      - application/json
      description: 获取指定时间范围内各模块的操作次数统计
      parameters:
      - description: 开始时间，格式：2006-01-02
        in: query
        name: start_time
        type: string
      - description: 结束时间，格式：2006-01-02
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_interfaces_http_handler.AuditModuleStatItem'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取模块统计
      tags:
      - 审计日志
  /audit/stats:
    get:
      consumes:
      - application/json
      description: 获取指定时间范围内的审计日志统计信息，包括总数、今日数量、登录次数、操作次数、错误次数等
      parameters:
      - description: 开始时间，格式：2006-01-02
        in: query
        name: start_time
        type: string
      - description: 结束时间，格式：2006-01-02
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_internal_domain_entity.AuditLogStats'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取审计日志统计
      tags:
      - 审计日志
  /audit/user-activity/{userId}:
    get:
      consumes:
      - application/json
      description: 获取指定用户在指定天数内的活动统计，按日期展示操作次数
      parameters:
      - description: 用户ID
        in: path
        name: userId
        required: true
        type: string
      - default: 7
        description: 天数范围，默认7天，最大90天
        in: query
        maximum: 90
        minimum: 1
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_interfaces_http_handler.AuditUserActivityItem'
                  type: array
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/github_com_Snowitty-Re_CNtunyuan_pkg_response.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取用户活动统计
      tags:
      - 审计日志
...
securityDefinitions:
  Bearer:
    description: 请输入 JWT Token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"